partial revert of [18217]. Inserting one vehicle per second for 70 years is probably enough
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@18231 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/vehicle/SUMOVehicleParameter.h
@@ -437,9 +437,9 @@ public:
     /// @{
 
     /// @brief The number of times the vehicle shall be repeatedly inserted
-    int_fast64_t repetitionNumber;
+    int repetitionNumber;
     /// @brief The number of times the vehicle was already inserted
-    int_fast64_t repetitionsDone;
+    int repetitionsDone;
     /// @brief The time offset between vehicle reinsertions
 #ifdef HAVE_SUBSECOND_TIMESTEPS
     SUMOTime repetitionOffset;
